fix: implement async synchronization using mutexes
Previous implementation tried to overcome synchronization problems
implicitly using serialization of async requests. Implementation was
flawed in that it stopped working when a curl request failed and
serialization was limited to calls to the same function.

This implementation makes synchronization explicit using
self-implemented mutexes. This commit also changes the async callback
signature to let the async module know of errors in a coroutine to
releases all mutexes this coroutine owns.