fix(engines)!: read default_engine from main table
The docs said that the key `default_engine` in the main table can be
used to configure the default engine to use. However the example in the
README sets this key in the engines table and the implementation reads
it from there, too. The docs describe the intended way so this. This fix
might break some existing configs but it is an easy fix. Raised in #2.