fix(config): apply user configs with customized keybindings
Since the actions module might be loaded before `pantran.setup` is
invoked, user configs might not be applied correctly to modules which
are required eagerly in it. This commit thus introduces lazy requiring
of modules used in the actions module. Related to that, the README
recommended `package.loaded` to access actions which actually doesn't
work in the way it was shown. The README is thus updated accordingly.
(Fixes #9)